fixed: Some compile errors.

+++ ascension/ascension/graphics/font/line-relative-direction.hpp
@@ -49,7 +49,7 @@ namespace ascension {
 					LineRelativeDirection::LINE_UNDER, LineRelativeDirection::LINE_OVER,
 					LineRelativeDirection::LINE_RIGHT, LineRelativeDirection::LINE_LEFT
 				};
-				const std::size_t index = direction;
+				const std::size_t index = static_cast<std::size_t>(direction);
 				if(index >= std::extent<decltype(opposites)>::value)
 					throw UnknownValueException("direction");
 				return opposites[index];
